Accessor tests no longer have dependency on Ginkgo
The accessor tests now make sure that no dependency (on public header)
exists inside the accessor headers by only having the source directory
of Ginkgo as an include directory (and no longer linking against
Ginkgo).
Only one test (`reduced_row_major_ginkgo`) still depends on Ginkgo since
it uses `gko::half`.
